Extract Configuration out of Framework.kt into its own file
